Upgrade path to ombi v4 (#491)
* Remove opld service, remove comments, fix echos to terminal

* Remove old service and use packaged one

* Add nginx fix in case that's installed

* make nginx rms conditional, remove apt sources

* make upgrade just go to the v4 apt sources and force install

* no reason to not overwrite

* just overwrite the config no?

* remove escapes again

* numbers are hard man

* Remove useless udpate

* make sure ombi is started after install

* formatting

* "formatting"

* Add confirmation to undowngradeable scenario

* hush

* Empty commit, testing pipeline

* teenyweenie fixes

* Update scripts/nginx/ombi.sh

* Aded noticesd about the port ombi is using, switch properly to 5000

* Actually perform a database backup

* Moving back to port 3000 using overrides

* little prod to set up the user and all

+++ scripts/remove/ombi.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
-systemctl disable -q ombi
-systemctl stop -q ombi
-rm /etc/systemd/system/ombi.service
-rm -f /etc/nginx/apps/ombi.conf
-systemctl reload nginx
+systemctl disable -q --now ombi
+if [[ -f /install/.nginx.lock ]]; then
+	rm -f /etc/nginx/apps/ombi.conf
+	systemctl reload nginx
+fi
+
+rm /etc/apt/sources.list.d/ombi.list
 
 apt_remove ombi
